papiea-sdk/python/e2e_tests/requirements.txt
aiohttp==3.6.2
asn1crypto==0.24.0
async-timeout==3.0.1
attrs==19.3.0
certifi==2020.6.20
chardet==3.0.4
cryptography==2.1.4
idna==2.10
inflect==4.1.0
keyring==10.6.0
keyrings.alt==3.0
more-itertools==8.4.0
multidict==4.7.6
packaging==20.4
pluggy==0.13.1
py==1.9.0
pycrypto==2.6.1
pygobject==3.26.1
pyparsing==2.4.7
pytest==5.4.3
pytest-asyncio==0.14.0
pyxdg==0.25
PyYAML==5.3.1
requests==2.24.0
SecretStorage==2.3.1
six==1.15.0
urllib3==1.25.10
wcwidth==0.2.5
yarl==1.4.2
